+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+## Added
+- Athena and Dionysus boon lists
 
 ## [0.0.1] - 2024-06-19